Create README.md
These examples are based on either the Linux Foundation or Google's courses available through Coursera (can't remember which). I was trying to replicate merge conflicts to make better sense of them, and a conflict I thought would appear in one of these examples did not end up happening because because of a fast forward merge. These script files create various text files and modify them. It's a small example, and I was mostly playing around here.